Added basic YAML parsing
A good start at parsing logic for YAML. Of course everything seems more
verbose (to the average Python developer) due to all of the type hints,
but it will make for a more reliable codebase that's easier to work
with. For example, types of YAML keys are now validated using
isinstance, a big step from the old "trust me bro" method in previous
versions of bpy-build D:

Based on limited testing, YAML parsing seems to work fine, we're getting
closer to an alpha build of the rewrite, perhaps in late August or early
September (maybe a birthday release?).