Renamed bpy-addon-build utility to bab
bpy-addon-build is 15 characters long, which is annoying to type in the
terminal. Even with tab complete, it still makes sense to reduce the
length of the name.

The name choosen is bab, which is derived from the first letters of each
word in the original name. Genius I know.

This does technically qualify as a breaking change, but who cares when
it means typing only 3 characters?