Added proper type checking with errors
Previously we had some type checking but it would only result in
BuildContext not being created. Now we have proper type checking with
error messages.

Due to isinstance not supporting generics, install_versions and actions
must be checked using validators (a feature in attrs). Not a big deal,
but it does mean error messages aren't as clean as they could be.