Added try-except loop around BuildContext creation
When BuildContext is being created, it also runs validators to check if
certain pieces of data are correct. If they aren't, then a ValueError is
thrown. This is better then nothing, but results in very unclean error
messages.

The try-except loop fixes that by printing just the error without the
traceback.