feat: Added bab-create command
bab-create is a new command that creates a basic BpyBuild project, with
the proper directories and configuration!

Right now it's extremely barebones, but in the future, we'll have
BpyBuild generate an example addon (just a "Hello World" panel, nothing
fancy) for the developer to use. We could even integrate it Git and also
create a git repository automatically!