feat: Add config options for Extensions
Blender 4.2 recently introduced Extensions to replace the old legacy
addons. As a starting point to support Extensions, this patch adds the
needed config options to support Extensions. These new config options
are:
- build_extension [bool]: the switch to enable extension building
- extension_settings::blender_binary [str]: the path to the Blender
  binary needed to build extensions (4.2+)
- extension_settings::build_legacy [bool]: the switch to build a legacy
  addon alongside the extension
- extension_settings::build_name [str]: an optional alternative name for
  extensions. By default however, we plan to do[build_name]-extension

It should be noted that this patch exclusively adds new config options,
the actual building of extensions is not supported yet. In addition, the
current roadmap for extension support involves the developer passing in
the direct path to the Blender binary, which isn't ideal for versions of
Blender from the Microsoft Store, Snap repo, and Flathub. Perhaps we may
implement our own extension builder.

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>