feat: Add extension_settings::remove_bl_info
This is intended for developers building both an extension and legacy
addon. The idea is that BpyBuild will be able to remove the `bl_info`
dictionary when building an addon as an extension, if and only if the
developer also has extension_settings::build_legacy enabled (otherwise
what's the point?).

This feature is planned to be opt-in, as it involves modifying an
addon's source code. However, perhaps at some point we could possibly
make it the default for building extensions with
extension_settings::build_legacy enabled

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>