feat: Add output_name and output_settings, deprecate build_name
One issue that exists with the build name system in BpyBuild is the fact
that it's not configurable with alternate outputs. BpyBuild 0.5 sort of
makes up for this with the `_legacy` suffix, but nothing configurable
for end users.

To make output names easier to configure, we've deprecated build_name
and replaced it with output_name, which would be a string with
formatting options. In addition, output_settings would allow for further
configuration.

Nothing is final yet, but the current plan is to deprecate build_name
but still accept it for 0.5.X, then remove it in 0.6.

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>