feat: Fully integrate output_name into BpyBuild
This allows for significantly more flexibility when creating multiple
builds of an addon, especially with extension+legacy builds or builds
that involve compiled components.

This only supports the `build_type` an `os` variables that can be used
in `output_name`, no `arch` or anything like that. In addition, there is
not yet support for custom variables that could be implemented using
actions, although that is planned as something to support later.

For operating systems, we split `linux` and `posix`, due to differences
between the 2 platforms. For our purposes, `posix` is FreeBSD, NetBSD, and
OpenBSD.

CPU architecture is not yet implemented because there's no end to the
rabbit hole that is `uname -m`, which is used on OSX, Linux, and the BSDs,
meanwhile Windows could give different results depending on if the
Python interpreter is running in something like Cygwin (which uses
`uname -m`) or Windows itself (which is completely different). We do
eventually plan to implement this, so for now, we'll just print a
warning saying support is not yet implemented.

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>