feat: Parse and validate subactions, and validate depends_on
Subactions are a new concept of actions that are ran directly after
another action. This is mainly to reduce issues with creating large
actions that do mulitple tasks (and really should be mulitple actions).

`depends_on` allows defining dependencies for actions. It was added a
while back, but validating that the actions in it exist wasn't enforced
until now. So far it isn't a used variable either.

Signed-off-by: Maryam Sheikh <mahid@standingpad.org>