feat: Run actions defined as subactions to an action
When constructing the API object, we first go through all of the actions
defined in the CLI, and see if they have any subactions. If they do, we
then add the subactions after the parent action. As far as the rest of
BpyBuild is concerned, the user had defined those actions to be ran in
the CLI.

Signed-off-by: Maryam Sheikh <mahid@standingpad.org>