feat: Implement dynamic_name hook and action-defined variables
Originally the plan was to save this for a later date, however in
working out how to implement the `version` variable, I realized that the
best route would be through an action with a custom name hook. As suc20
I figured it would be best to just create the whole action-defined
variables concept now and implement the version variable using it.

There's 2 parts to creating action-defined variables: first, they must
be declared in `output_settings` with the `@dynamic` or
`@dynamic_required` keywords, as follows:

```yaml
output_settings:
  dynamic_var: @dynamic
  dynamic_var_2: @dynamic_required
```

The `@dynamic` keyword is reserved, and signifies that a variable is
dynamically defined in an action. However, it should be noted that the
variable is not guaranteed to be defined. Once fallbacks are
implemented, BpyBuild will enforce dynamic values to have a fallback
when used in `output_name`.

Alternatively, the `@dynamic_required` keyword can be used, which
BpyBuild enforces the existance of a dynamic variable after all name
related hooks are ran. This is what the `version` variable will use.

The second part is the new `dynamic_name` hook, which is called after
the `pre_build` hook, and bfore the `main` hook. This hook is of type
`Callable[[BabContext], Optional[Union[list[BabVariableDef], BabWarning, BabError]]`,
in which a list of variable definitions is returned. BpyBuild then uses
the returned list to add to the dictionary of defined variables, which
in turn is used for the final name formatting.

Signed-off-by: Maryam Sheikh <mahid@standingpad.org>