feat: Add fallbacks and concatenation for output_name
Sometimes multiple variables may need to be used with a fallback, like
falling back to a different OS depending on what variable is defined.
This adds the fallback syntax of `{var1|var2}`, where BpyBuild will
either interpret the expression as `{var1}` (if `var1` is defined) or
`{var2}` (if `var2` is defined and `var1` isn't).

If an expression contains 2 variables that aren't defined, then by
default, BpyBuild will throw an error. If a user wants to simply have
nothing there, then the expression `{var1|var2|""}` can be used, where
BpyBuild uses the empty string if both `var1` and `var2` are not
defined.

This also adds the ability to concatenate a string with a variable while
also using an expression. For example, in the case of {y + "_build"},
the variable `y` will be combined with the string `"build"`. This is
especially useful when combined with fallbacks, such as `{x | y +
"_build"}`.

It should be noted that the left and right side of the | operator are
evaluated beforehand, since at runtime it expressions are converted to
something akin to RPN (using an adapted form of the shunting yard
algorithm found in many basic calculators). If an expression would be
undefined, internally it's evaluated as `UNDEFINED`.

Signed-off-by: Maryam Sheikh <mahid@standingpad.org>